# backend/docs/swagger.yaml
basePath: /
definitions:
...
  api.BoolResponse:
    properties:
      ok:
        type: boolean
    type: object
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  db.AgentPresenceRow:
    properties:
      break_since:
        type: string
      email:
        type: string
      full_name:
        type: string
      idle_since:
        type: string
      last_seen_at:
        type: string
      manager_id:
        type: integer
      on_call_since:
        type: string
      role:
        type: string
      status:
        type: string
      total_idle_time_s:
        type: integer
      total_talk_time_s:
        type: integer
      user_id:
        type: integer
    type: object
...
host: localhost:8011
info:
  contact:
    email: support@callified.ai
    name: Callified Support
  description: AI-powered outbound calling, WhatsApp CRM & receptionist platform.
  termsOfService: http://callified.ai/terms
  title: Callified Backend API
  version: "1.0"
paths:
...
  /api/presence:
    get:
      description: Returns live status for every agent in the org. Admin only.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.AgentPresenceRow'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List agent presence
      tags:
      - presence
  /api/presence/break:
    post:
      description: Convenience endpoint for the agent dashboard to mark themselves
        on break.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BoolResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Set agent on break
      tags:
      - presence
  /api/presence/heartbeat:
    post:
      consumes:
      - application/json
      description: Agents ping this endpoint periodically to report their current
        status.
      parameters:
      - description: 'status: idle | break | on_call'
        in: body
        name: body
        schema:
          properties:
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BoolResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Agent presence heartbeat
      tags:
      - presence
  /api/presence/idle:
    post:
      description: Marks the current agent as idle (available).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BoolResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Clear break / set agent idle
      tags:
      - presence
...
  /api/sse/notifications:
    get:
      description: Server-sent event stream for in-app notifications. Authenticate
        via ?ticket=<sse-ticket> or Authorization header.
      parameters:
      - description: Short-lived SSE ticket
        in: query
        name: ticket
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: 'event: notification\ndata: {...}\n\n'
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Notification stream (SSE)
      tags:
      - sse
  /api/sse/presence:
    get:
      description: Server-sent event stream of agent presence changes for the admin
        dashboard.
      parameters:
      - description: Short-lived SSE ticket
        in: query
        name: ticket
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: 'data: {...}\n\n'
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Presence SSE stream
      tags:
      - sse
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter: **Bearer {JWT token}**. Obtain a token from POST /api/auth/login
      or POST /api/auth/signup.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"